requirements.txt
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai==1.14.3
python-dotenv==1.0.1
sqlite3  
requests

